requirements.txt
pyspark==3.2.4
cx_Oracle==8.3.0
PyYAML==6.0.1

# 테스트 관련
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

